Remove CI resource class workaround due to grpc issue
While being installed, the grpc gem compiles a bunch of native code.

In #1791 I documented an issue we were running into in Circle CI where
the CI worker was running out of memory because grpc was spawning
too many parallel processes for compilation.

I brought the issue up with upstream in
https://github.com/grpc/grpc/issues/28244 and in
https://github.com/grpc/grpc/pull/28250 a new flag was added to
enable us to limit the number of processes used.

So now we can remove our workaround of bumping the resource class
put in place in #1790 and instead use the `GRPC_RUBY_BUILD_PROCS`
flag to tell grpc how many processes to use.

Fixes #1791